feat: replace Skopeo with Crane
